Removed the old roadmap-items-grid component that was using ag-grid.
